Refactor metadata validation errors to use ValidationError struct
